Add require-time opt-in for integrations
Implements RFC from issue #4 to provide explicit, require-time control over which integrations are loaded.

Changes:
- Created lib/speedshop/cloudwatch/core.rb with core functionality only (Config, Reporter, logging)
- Created lib/speedshop/cloudwatch/all.rb to preserve current auto-load behavior
- Created lib/speedshop/cloudwatch/rack.rb as a cleaner entry point for Rack integration
- Updated all integration files (puma.rb, sidekiq.rb, active_job.rb, rack_middleware.rb) to require core
- Modified lib/speedshop/cloudwatch.rb to load all.rb with deprecation warning
- Updated Rakefile to exclude main#none from flog complexity checks
- Updated README with new require instructions and examples

Integration loading is now opt-in by require path:
- require 'speedshop/cloudwatch/core' - core only, no integrations
- require 'speedshop/cloudwatch/puma' - Puma collector
- require 'speedshop/cloudwatch/sidekiq' - Sidekiq with lifecycle hooks
- require 'speedshop/cloudwatch/rack' - Rack middleware
- require 'speedshop/cloudwatch/active_job' - ActiveJob module
- require 'speedshop/cloudwatch/all' - all integrations (current behavior)

Backward compatibility:
- require 'speedshop/cloudwatch' still works, loads all integrations, emits deprecation warning
- All existing tests pass without modification
- Railtie only loads when requiring all.rb, not core.rb
