+++ lib/speedshop/cloudwatch/rack.rb
@@ -1,0 +1,3 @@
+# frozen_string_literal: true
+
+require "speedshop/cloudwatch/rack_middleware"
